requirements.txt
annotated-types==0.7.0
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
decorator==4.4.2
distro==1.9.0
fal_client==0.4.1
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
Jinja2==3.1.4
jiter==0.5.0
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai==1.47.1
openai-whisper==20231117
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
proglog==0.1.10
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
regex==2024.9.11
replicate==0.33.0
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.1
socksio==1.0.0
sympy==1.13.3
tiktoken==0.7.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
shortcap==1.0.5

